pyproject.toml
[tool.poetry]
name = "infrastructure"
version = "0.1.0"
description = ""
authors = ["cch0807 <ckdgus11021@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.103.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
flake8 = "^6.1.0"
isort = "^5.12.0"
black = "^23.9.1"
alembic = "^1.12.0"
sqlalchemy = "^2.0.21"
pytz = "^2023.3.post1"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
